Hu: reorganize bash<->C++ program interface
Now we invoke program only once with list of images and get JSON file with array of objects,
each representing Hu moments and a learning label. Previously we had to invoke the program from bash several times,
each program call apending and object to JSON array in file which was was unfeasible.